// src/app/logs/models/buff-id.enum.ts
export enum BuffId {
  BERSERKING = 26297,
  BLOODLUST = 2825,
  BREATH_HASTE = 45856,
  DRUMS_OF_BATTLE = 35476,
  GREATER_DRUMS_OF_BATTLE = 351355,
  FEL_INFUSION = 40396,
  FOCUS = 18803,
  HASTE = 28507,
  HEROISM = 32182,
  MOONKIN_AURA = 24907,
  POWER_INFUSION = 10060,
  QUAGS_EYE = 33370,
  RETRIBUTION_AURA = 27150,
  WRATH_OF_AIR = 3738
}